// src/api/sqlmodule/model/attemptResultSnapshotState.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * SQLModule API
 * API SQL-тренажёра (модуль Scoodle).
 * OpenAPI spec version: v1
 */

export type AttemptResultSnapshotState = typeof AttemptResultSnapshotState[keyof typeof AttemptResultSnapshotState];


export const AttemptResultSnapshotState = {
  NotStored: 'NotStored',
  Available: 'Available',
  NotProduced: 'NotProduced',
  Expired: 'Expired',
} as const;